+++ packages/dockest/src/run/waitForServices/index.spec.ts
@@ -1,17 +1,17 @@
 import { EventEmitter } from 'events'
-import { waitForServices } from './index'
+import { waitForServices } from '.'
 import { checkConnection } from './checkConnection'
 import { runReadinessCheck } from './runReadinessCheck'
 // import { fixRunnerHostAccessOnLinux } from './fixRunnerHostAccessOnLinux'
+import { dockerComposeUp } from './dockerComposeUp'
 import { resolveContainerId } from './resolveContainerId'
 import { runRunnerCommands } from './runRunnerCommands'
-import { dockerComposeUp } from './dockerComposeUp'
-import { joinBridgeNetwork } from '../../utils/network/joinBridgeNetwork'
-import { sleepWithLog } from '../../utils/sleepWithLog'
+import { sleepWithLog } from '../..'
+import { createRunner } from '../../test-utils'
+import { getOpts } from '../../utils/get-opts'
 import { bridgeNetworkExists } from '../../utils/network/bridgeNetworkExists'
 import { createBridgeNetwork } from '../../utils/network/createBridgeNetwork'
-import { createRunner } from '../../test-utils'
-import { getOpts } from '../../utils/getOpts'
+import { joinBridgeNetwork } from '../../utils/network/joinBridgeNetwork'
 import { LogWriter } from '../log-writer'
 
 jest.mock('./checkConnection')
