+++ packages/dockest/src/utils/hash-code.spec.ts
@@ -1,4 +1,4 @@
-import { hashCode } from './hashCode'
+import { hashCode } from './hash-code'
 
 describe('hashCode', () => {
   describe('happy', () => {
