+++ packages/dockest/src/utils/hash-code.ts
