+++ packages/dockest/src/utils/network/createBridgeNetwork.ts
@@ -1,5 +1,5 @@
 import { BRIDGE_NETWORK_NAME } from '../../constants'
-import { execaWrapper } from '../execaWrapper'
+import { execaWrapper } from '../execa-wrapper'
 
 export const createBridgeNetwork = async () => {
   const command = `docker network create \
