+++ packages/dockest/src/utils/network/joinBridgeNetwork.ts
@@ -1,5 +1,5 @@
 import { BRIDGE_NETWORK_NAME } from '../../constants'
-import { execaWrapper } from '../execaWrapper'
+import { execaWrapper } from '../execa-wrapper'
 
 export const joinBridgeNetwork = async ({ containerId, alias }: { containerId: string; alias: string }) => {
   const command = `docker network connect \
