+++ packages/dockest/src/utils/network/leaveBridgeNetwork.ts
@@ -1,5 +1,5 @@
 import { BRIDGE_NETWORK_NAME } from '../../constants'
-import { execaWrapper } from '../execaWrapper'
+import { execaWrapper } from '../execa-wrapper'
 
 export const leaveBridgeNetwork = async ({ containerId }: { containerId: string }) => {
   const command = `docker network disconnect \
