+++ packages/dockest/src/utils/select-port-mapping.spec.ts
@@ -1,4 +1,4 @@
-import { selectPortMapping } from './selectPortMapping'
+import { selectPortMapping } from './select-port-mapping'
 
 describe('selectPortMapping', () => {
   it.each([
